refactor(patchers): move globals out of patchers
