+++ src/shared/globals/OC/dialogs.js
